fix queue testcode
